# api.swagger.yaml
swagger: "2.0"
info:
  title: Blog Server.
  description: Entry point for the application.
  version: 1.0.0
  contact:
    name: HardDie
    url: https://github.com/HardDie
    email: oleg1995sysoev@yandex.ru
...
basePath: /
schemes:
  - https
  - http
consumes:
  - application/json
produces:
  - application/json
paths:
...
  /api/v1/user/password:
    put:
      summary: Updating the password for a user
      operationId: User_Password
      responses:
        "200":
          description: A successful response.
          schema:
            type: object
            properties: {}
        default:
          description: An unexpected error response.
          schema:
            $ref: '#/definitions/rpcStatus'
      parameters:
        - name: body
          in: body
          required: true
          schema:
            $ref: '#/definitions/gatewayPasswordRequest'
      tags:
        - User
  /api/v1/user/profile:
    put:
      summary: Updating user information
      operationId: User_Profile
      responses:
        "200":
          description: A successful response.
          schema:
            type: object
            properties: {}
        default:
          description: An unexpected error response.
          schema:
            $ref: '#/definitions/rpcStatus'
      parameters:
        - name: body
          in: body
          required: true
          schema:
            $ref: '#/definitions/gatewayProfileRequest'
      tags:
        - User
  /api/v1/user/{id}:
    get:
      summary: Getting information about a user by ID
      operationId: User_Get
      responses:
        "200":
          description: A successful response.
          schema:
            $ref: '#/definitions/gatewayGetResponse'
        default:
          description: An unexpected error response.
          schema:
            $ref: '#/definitions/rpcStatus'
      parameters:
        - name: id
          in: path
          required: true
          type: string
          format: int64
      tags:
        - User
definitions:
...
  gatewayGetResponse:
    type: object
    properties:
      data:
        $ref: '#/definitions/gatewayPrivateUserObject'
...
  gatewayPasswordRequest:
    type: object
    properties:
      oldPassword:
        type: string
      newPassword:
        type: string
...
  gatewayPrivateUserObject:
    type: object
    properties:
      id:
        type: string
        format: int64
      username:
        type: string
      displayedName:
        type: string
      email:
        type: string
      invitedByUserId:
        type: string
        format: int64
      createdAt:
        type: string
        format: date-time
  gatewayProfileRequest:
    type: object
    properties:
      displayedName:
        type: string
      email:
        type: string
...
  protobufAny:
    type: object
    properties:
      '@type':
        type: string
    additionalProperties: {}
  rpcStatus:
    type: object
    properties:
      code:
        type: integer
        format: int32
      message:
        type: string
      details:
        type: array
        items:
          type: object
          $ref: '#/definitions/protobufAny'
securityDefinitions:
  bearer:
    type: apiKey
    description: 'Authentication token, prefixed by Bearer: Bearer <token>'
    name: Authorization
    in: header
security:
  - bearer: []